fix: refactor and move various functions
