+++ ts/themes/SessionTheme.tsx
@@ -2,8 +2,8 @@ import { ipcRenderer } from 'electron';
 import React from 'react';
 import { createGlobalStyle } from 'styled-components';
 import { switchThemeTo } from './switchTheme';
-import { ThemeStateType } from './constants/colors';
 import { classicDark } from './classicDark';
+import { getOppositeTheme, isThemeMismatched } from '../util/theme';
 import { declareCSSVariables, THEME_GLOBALS } from './globals';
 
 // Defaults to Classic Dark theme
@@ -21,28 +21,15 @@ export const SessionTheme = ({ children }: { children: any }) => (
   </>
 );
 
-export function getOppositeTheme(themeName: string) {
-  if (themeName.includes('dark')) {
-    return themeName.replace('dark', 'light');
-  }
-  if (themeName.includes('light')) {
-    return themeName.replace('light', 'dark');
-  }
-  // If neither 'dark' nor 'light' is in the theme name, return the original theme name.
-  return themeName;
-}
-
-export async function checkThemeCongruency(): Promise<boolean> {
+export async function ensureThemeConsistency(): Promise<boolean> {
   const theme = window.Events.getThemeSetting();
 
   return new Promise(resolve => {
     ipcRenderer.send('get-native-theme');
     ipcRenderer.once('send-native-theme', (_, shouldUseDarkColors) => {
-      const isMismatchedTheme =
-        (shouldUseDarkColors && theme.includes('light')) ||
-        (!shouldUseDarkColors && theme.includes('dark'));
+      const isMismatchedTheme = isThemeMismatched(theme, shouldUseDarkColors);
       if (isMismatchedTheme) {
-        const newTheme = getOppositeTheme(theme) as ThemeStateType;
+        const newTheme = getOppositeTheme(theme);
         void switchThemeTo({
           theme: newTheme,
           mainWindow: true,
